Convert to environment based detectors
